bincrafters/requirements.txt
conan_package_tools>=0.31.1
PyYAML>=3.11, <6.0
